Mobile First
This is the mobile version of my boutique website.  Tried to keep it
clean and simple and focus on the most important things for an online
shopper. Having the ability to see/access items easily and see shopping
cart. In the interest of the company the shopper can easily share items
they find and like them on their favorite social media site.